Avoid blocking in BigQuerySplitManager
